# services/meta-center/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 元数据服务API
  description: |
    元数据驱动的低代码平台核心服务，提供统一的元数据管理、动态数据操作、业务规则引擎、权限控制等能力。
    
    ## 核心功能
    - **元数据管理**: 支持元数据对象、属性、关系、规则的全生命周期管理
    - **动态数据服务**: 基于元数据的动态CRUD操作，支持表单渲染和数据校验
    - **业务规则引擎**: 动态配置和执行业务规则，支持多种表达式类型
    - **权限与安全**: 细粒度权限控制和数据脱敏
    - **多租户支持**: 租户级别的数据和配置隔离
    
    ## 版本与兼容性
    - API版本遵循语义化版本规范
    - 向后兼容性保证，废弃接口将提前通知
    - 支持接口版本协商和灰度发布
  version: 1.0.0
  contact:
    name: 元数据服务团队
    email: metadata-team@aixone.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.aixone.com/meta-center/v1
    description: 生产环境
  - url: https://api-staging.aixone.com/meta-center/v1
    description: 测试环境
  - url: http://localhost:8080/meta-center/v1
    description: 本地开发环境
security:
  - BearerAuth: []
  - TenantHeader: []
paths:
...
  /rules:
    get:
      tags:
        - 业务规则
      summary: 查询业务规则列表
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: ruleType
          in: query
          description: 规则类型过滤
          schema:
            type: string
            enum: [validation, automation, workflow, calculation]
        - name: scope
          in: query
          description: 作用范围过滤
          schema:
            type: string
        - name: enabled
          in: query
          description: 启用状态过滤
          schema:
            type: boolean
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedRuleResponse'
    
    post:
      tags:
        - 业务规则
      summary: 创建业务规则
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
  /rules/{ruleId}:
    get:
      tags:
        - 业务规则
      summary: 获取业务规则详情
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleDetailResponse'
    
    put:
      tags:
        - 业务规则
      summary: 更新业务规则
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuleRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
    
    delete:
      tags:
        - 业务规则
      summary: 删除业务规则
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
  /rules/{ruleId}/execute:
    post:
      tags:
        - 业务规则
      summary: 执行业务规则
      description: 手动触发规则执行，主要用于测试和调试
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRuleRequest'
      responses:
        '200':
          description: 执行完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleExecutionResult'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证token
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: 租户标识
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码，从0开始
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: 每页大小
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    MetaRule:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: string
          description: 规则ID
        name:
          type: string
          description: 规则名称
        type:
          type: string
          enum: [validation, automation, workflow, calculation, permission]
          description: 规则类型
        scope:
          type: string
          description: 作用范围
        expression:
          type: string
          description: 规则表达式
        dsl:
          type: string
          description: DSL表达式
        condition:
          type: string
          description: 触发条件
        action:
          type: string
          description: 执行动作
        message:
          type: string
          description: 提示信息
        enabled:
          type: boolean
          default: true
          description: 是否启用
        priority:
          type: integer
          default: 0
          description: 优先级
        trigger:
          type: string
          enum: [onCreate, onUpdate, onDelete, onSave, onValidate, manual]
          description: 触发时机
        group:
          type: string
          description: 规则组
        tags:
          type: array
          items:
            type: string
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRuleRequest:
      type: object
      required: [name, type, expression]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        type:
          type: string
          enum: [validation, automation, workflow, calculation, permission]
        scope:
          type: string
        expression:
          type: string
        condition:
          type: string
        action:
          type: string
        message:
          type: string
        priority:
          type: integer
        trigger:
          type: string
          enum: [onCreate, onUpdate, onDelete, onSave, onValidate, manual]
        group:
          type: string
        tags:
          type: array
          items:
            type: string
    UpdateRuleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        expression:
          type: string
        condition:
          type: string
        action:
          type: string
        message:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        tags:
          type: array
          items:
            type: string
    RuleResponse:
      $ref: '#/components/schemas/MetaRule'
    RuleDetailResponse:
      allOf:
        - $ref: '#/components/schemas/MetaRule'
        - type: object
          properties:
            executionHistory:
              type: array
              items:
                $ref: '#/components/schemas/RuleExecution'
            performance:
              type: object
              properties:
                avgExecutionTime:
                  type: number
                  description: 平均执行时间（毫秒）
                totalExecutions:
                  type: integer
                  description: 总执行次数
                successRate:
                  type: number
                  description: 成功率
    PagedRuleResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RuleResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
    ExecuteRuleRequest:
      type: object
      required: [context]
      properties:
        context:
          type: object
          additionalProperties: true
          description: 执行上下文
        dryRun:
          type: boolean
          default: false
          description: 是否为试运行
    RuleExecutionResult:
      type: object
      properties:
        ruleId:
          type: string
        success:
          type: boolean
          description: 是否执行成功
        result:
          description: 执行结果
        message:
          type: string
          description: 结果信息
        executionTime:
          type: number
          description: 执行耗时（毫秒）
        timestamp:
          type: string
          format: date-time
    RuleExecution:
      type: object
      properties:
        id:
          type: string
        ruleId:
          type: string
        context:
          type: object
          additionalProperties: true
        result:
          description: 执行结果
        success:
          type: boolean
        message:
          type: string
        executionTime:
          type: number
        timestamp:
          type: string
          format: date-time
...